move parser to lib
